+++ src/Monomer/Hagrid.hs
@@ -19,6 +19,7 @@ module Monomer.Hagrid
     ScrollToRowCallback,
 
     -- * Configuration options
+    estimatedItemHeight,
     initialSort,
 
     -- * Hagrid constructors
@@ -43,7 +44,6 @@ import Control.Monad as X (forM_)
 import Data.Data (Typeable)
 import Data.Default.Class (Default, def)
 import Data.Foldable (foldl')
-import qualified Data.List as List
 import Data.List.Index (indexed, izipWith, modifyAt)
 import Data.Maybe (catMaybes, fromJust, isNothing, maybeToList)
 import Data.Maybe as X (fromMaybe)
@@ -59,18 +59,38 @@ import Monomer.Widgets.Container
 import Monomer.Widgets.Single
 
 -- | Configuration options for Hagrid widgets.
-newtype HagridCfg s e = HagridCfg
-  { cfgInitialSort :: Maybe (Int, SortDirection)
+data HagridCfg s e = HagridCfg
+  { cfgEstimatedItemHeight :: Maybe Double,
+    cfgInitialSort :: Maybe (Int, SortDirection)
   }
 
 instance Default (HagridCfg s e) where
-  def = HagridCfg {cfgInitialSort = Nothing}
+  def =
+    HagridCfg
+      { cfgEstimatedItemHeight = Nothing,
+        cfgInitialSort = Nothing
+      }
 
 instance Semigroup (HagridCfg s e) where
-  c1 <> c2 = HagridCfg {cfgInitialSort = c2.cfgInitialSort <|> c1.cfgInitialSort}
+  c1 <> c2 =
+    HagridCfg
+      { cfgEstimatedItemHeight = c2.cfgEstimatedItemHeight <|> c1.cfgEstimatedItemHeight,
+        cfgInitialSort = c2.cfgInitialSort <|> c1.cfgInitialSort
+      }
 
 instance Monoid (HagridCfg s e) where
   mempty = def
+
+-- | Configures the estimated item height. This should be the average row height you expect
+-- in your grid (including padding). This is used to show scrollbar size and position when
+-- there are lots of rows and so not all the rows have been "inflated" into widgets. More
+-- accurate values will improve performance and scrollbar position accuracy.
+--
+-- The default value is 40, which is roughly the height of a single line of text with the default
+-- column padding (10).
+estimatedItemHeight :: Double -> HagridCfg s e
+estimatedItemHeight h =
+  def {cfgEstimatedItemHeight = Just h}
 
 -- | Configures the initial sort column and direction.
 initialSort ::
@@ -80,7 +100,7 @@ initialSort ::
   SortDirection ->
   HagridCfg s e
 initialSort column direction =
-  HagridCfg {cfgInitialSort = Just (column, direction)}
+  def {cfgInitialSort = Just (column, direction)}
 
 -- | A column definition.
 data Column e a = Column
@@ -99,9 +119,9 @@ data Column e a = Column
     initialWidth :: Int,
     -- | The minimum allowed width of the column, in pixels.
     minWidth :: Int,
-    -- | The padding to the left and right of the widget in each cell of the column, in pixels.
+    -- | The padding to the left and right of the widget in each cell of the column, in pixels (the default is 10).
     paddingW :: Double,
-    -- | The padding above and below the widget in each cell in the column, in pixels.
+    -- | The padding above and below the widget in each cell in the column, in pixels (the default is 10).
     paddingH :: Double,
     -- | An optional event to emit when a user has finished resizing the column. The function receives the new width in pixels.
     resizeHandler :: Maybe (Int -> e),
@@ -124,7 +144,7 @@ data ColumnFooterWidget e a
     NoFooterWidget
   | -- | Create a footer widget. The function receives the items in their current sort
     -- order, and also along with each item it's original (unsorted) index.
-    CustomFooterWidget (forall s. WidgetModel s => [ItemWithIndex a] -> WidgetNode s e)
+    CustomFooterWidget (forall s. WidgetModel s => Seq (ItemWithIndex a) -> WidgetNode s e)
 
 -- | How to align the widget within each cell of a column.
 data ColumnAlign
@@ -152,7 +172,7 @@ type ItemWithIndex a = (a, Int)
 type ScrollToRowCallback a =
   -- | The items in the grid, in the originally provided order, along with each item's index
   -- in the current grid order.
-  [ItemWithIndex a] ->
+  Seq (ItemWithIndex a) ->
   -- | The row to scroll to, as an index into the sorted items (e.g. 0 is always the first row
   -- in the grid, regardless of the current order). 'Nothing' will cancel the scroll.
   Maybe Int
@@ -163,12 +183,14 @@ data HagridEvent ep
   | ResizeColumn Int Int
   | ResizeColumnFinished Int
   | forall a. Typeable a => ScrollToRow (ScrollToRowCallback a)
+  | ScrollToRect Rect
   | ParentEvent ep
 
 data HagridModel a = HagridModel
-  { sortedItems :: [ItemWithIndex a],
+  { sortedItems :: Seq (ItemWithIndex a),
     columns :: [ModelColumn],
-    sortColumn :: Maybe (Int, SortDirection)
+    sortColumn :: Maybe (Int, SortDirection),
+    mdlEstimatedItemHeight :: Double
   }
   deriving (Eq, Show)
 
@@ -193,8 +215,35 @@ data HeaderDragHandleState = HeaderDragHandleState
   }
   deriving (Eq, Show)
 
-newtype ContentPaneMessage a
-  = ContentPaneScrollToRow (ScrollToRowCallback a)
+data ContentPaneModel a = ContentPaneModel
+  { -- | The width of each column.
+    columnWidths :: [Int],
+    -- | The visible area (relative to the content pane).
+    visibleArea :: Rect,
+    -- | The index in items of the special row that we try and keep in the same position
+    -- (relative to the viewport) when items get resized.
+    fixedRowIndex :: Int,
+    -- | The Y position of the fixed row within the viewport. When items get resized, we try
+    -- and keep this value fixed: it should only change when the user scrolls the viewport itself.
+    fixedRowViewportOffset :: Double,
+    -- | How many items there are before the inflated items.
+    itemsBeforeInflated :: Int,
+    -- | The items that have been inflated into actual widgets.
+    inflatedItems :: Seq (ItemWithIndex a),
+    phase :: ContentPanePhase
+  }
+  deriving (Eq, Show)
+
+data ContentPanePhase
+  = ContentPaneIdle
+  | ContentPaneReinflating
+  deriving (Eq, Show)
+
+data ContentPaneEvent ep
+  = SetVisibleArea {visibleArea :: Rect}
+  | InnerResizeComplete
+  | ContentPaneParentEvent ep
+  | forall a. Typeable a => ContentPaneScrollToRow (ScrollToRowCallback a)
 
 -- | Creates a hagrid widget, using the default configuration.
 hagrid ::
@@ -203,7 +252,7 @@ hagrid ::
   -- | The definitions for each column in the grid.
   [Column e a] ->
   -- | The items for each row in the grid.
-  [a] ->
+  Seq a ->
   WidgetNode s e
 hagrid = hagrid_ def
 
@@ -215,7 +264,7 @@ hagrid_ ::
   -- | The definitions for each column in the grid.
   [Column e a] ->
   -- | The items for each row in the grid.
-  [a] ->
+  Seq a ->
   WidgetNode s e
 hagrid_ cfg columnDefs items = widget
   where
@@ -240,16 +289,21 @@ hagrid_ cfg columnDefs items = widget
             ]
         contentScroll =
           scroll_ [onChange ContentScrollChange] $
-            contentPane columnDefs model `nodeKey` contentPaneKey
+            contentPaneOuter columnDefs model `nodeKey` contentPaneOuterKey
 
     handleEvent :: EventHandler (HagridModel a) (HagridEvent e) sp e
     handleEvent wenv _node model = \case
-      ScrollToRow row ->
-        [Message (WidgetKey contentPaneKey) (ContentPaneScrollToRow row)]
-      ContentScrollChange ScrollStatus {scrollDeltaX} ->
+      ScrollToRow callback ->
+        [Message (WidgetKey contentPaneOuterKey) (ContentPaneScrollToRow callback :: ContentPaneEvent e)]
+      ScrollToRect rect ->
+        [Message (WidgetKey contentScrollKey) (ScrollTo rect)]
+      ContentScrollChange ScrollStatus {scrollDeltaX, scrollDeltaY, scrollVpSize} ->
         [ Message (WidgetKey headerPaneKey) (SetOffsetX scrollDeltaX),
+          Message (WidgetKey contentPaneOuterKey) (SetVisibleArea {visibleArea} :: ContentPaneEvent e),
           Message (WidgetKey footerPaneKey) (SetOffsetX scrollDeltaX)
         ]
+        where
+          visibleArea = Rect (-scrollDeltaX) (-scrollDeltaY) scrollVpSize._sW scrollVpSize._sH
       OrderByColumn colIndex -> result
         where
           Column {sortHandler, sortKey} = columnDefs !! colIndex
@@ -257,7 +311,7 @@ hagrid_ cfg columnDefs items = widget
             | Just (c, dir) <- model.sortColumn,
               c == colIndex =
                 let sortColumn = Just (colIndex, flipSortDirection dir)
-                    sortedItems = reverse model.sortedItems
+                    sortedItems = S.reverse model.sortedItems
                  in (sortColumn, sortedItems)
             | otherwise =
                 let sortColumn = Just (colIndex, SortAscending)
@@ -271,8 +325,7 @@ hagrid_ cfg columnDefs items = widget
       ResizeColumn colIndex newWidth ->
         [ Model (model {columns = modifyAt colIndex (\c -> c {currentWidth = newWidth}) model.columns}),
           Request (ResizeWidgets headerPaneId),
-          Request (ResizeWidgets footerPaneId),
-          Request (ResizeWidgets contentPaneId)
+          Request (ResizeWidgets footerPaneId)
         ]
       ResizeColumnFinished colIndex -> result
         where
@@ -285,7 +338,6 @@ hagrid_ cfg columnDefs items = widget
       where
         headerPaneId = fromJust (widgetIdFromKey wenv (WidgetKey headerPaneKey))
         footerPaneId = fromJust (widgetIdFromKey wenv (WidgetKey footerPaneKey))
-        contentPaneId = fromJust (widgetIdFromKey wenv (WidgetKey contentPaneKey))
 
     -- If the column names have not changed then preseve the column widths and sort
     -- order too, otherwise unrelated model changes will reset the column widths/sort.
@@ -562,161 +614,341 @@ headerDragHandle colIndex columnDef column = tree
 hScrollFudgeFactor :: Double
 hScrollFudgeFactor = 100
 
-contentPane ::
+-- | Composite wrapper to allow creating/removing child widgets during resize.
+contentPaneOuter ::
   forall a ep.
   (CompositeModel a, WidgetEvent ep) =>
   [Column ep a] ->
   HagridModel a ->
   WidgetNode (HagridModel a) (HagridEvent ep)
-contentPane columnDefs model = node
+contentPaneOuter columnDefs model =
+  compositeD_
+    "Hagrid.ContentPaneOuter"
+    (WidgetValue initialModel)
+    buildUI
+    handleEvent
+    [compositeMergeModel mergeModel]
+  where
+    initialModel =
+      ContentPaneModel
+        { columnWidths = currentWidth <$> model.columns,
+          visibleArea = Rect 0 0 0 0,
+          fixedRowIndex = 0,
+          fixedRowViewportOffset = 0,
+          itemsBeforeInflated = 0,
+          inflatedItems = mempty,
+          phase = ContentPaneIdle
+        }
+
+    mergeModel :: MergeModelHandler (ContentPaneModel a) (ContentPaneEvent e) s
+    mergeModel _wenv _parentModel oldModel newModel =
+      oldModel {columnWidths, fixedRowIndex, itemsBeforeInflated, inflatedItems}
+      where
+        columnWidths = newModel.columnWidths
+        fixedRowIndex = min (max 0 (length model.sortedItems - 1)) oldModel.fixedRowIndex
+        itemsBeforeInflated = min (length model.sortedItems) oldModel.itemsBeforeInflated
+        inflatedItems = takeAt itemsBeforeInflated (length oldModel.inflatedItems) model.sortedItems
+
+    buildUI _wenv cpModel =
+      contentPaneInner (S.fromList columnDefs) model cpModel `nodeKey` contentPaneInnerKey
+
+    handleEvent wenv node cpModel = \case
+      SetVisibleArea visibleArea -> result
+        where
+          result = case cpModel.phase of
+            ContentPaneIdle
+              | visibleAreaMoved && (startItemsMissing || endItemsMissing) ->
+                  let newModel =
+                        cpModel
+                          { visibleArea,
+                            fixedRowIndex,
+                            fixedRowViewportOffset,
+                            itemsBeforeInflated = fixedRowIndex,
+                            inflatedItems = mempty,
+                            phase = ContentPaneReinflating
+                          }
+                   in [Model newModel, resizeInnerRequest wenv]
+              | visibleAreaMoved ->
+                  let newModel = cpModel {visibleArea, fixedRowIndex, fixedRowViewportOffset}
+                   in [Model newModel]
+              | otherwise -> []
+            ContentPaneReinflating
+              | visibleAreaMoved ->
+                  let newModel = (cpModel :: ContentPaneModel a) {visibleArea}
+                   in [Model newModel]
+              | otherwise -> []
+
+          (fixedRowIndex, fixedRowViewportOffset) = fixedRow minVisibleY rowHeights model cpModel
+          (rowsStartY, rowHeights, rowsEndY) = rowPositions node
+
+          visibleAreaMoved = not (roundedRectEq visibleArea cpModel.visibleArea)
+          minVisibleY = visibleArea._rY
+          maxVisibleY = visibleArea._rY + visibleArea._rH
+
+          startItemsMissing = cpModel.itemsBeforeInflated > 0 && minVisibleY < rowsStartY
+          endItemsMissing = itemsAfterInflated model cpModel > 0 && maxVisibleY > rowsEndY
+      ContentPaneScrollToRow callback -> result
+        where
+          result
+            | Just typedCb <- cast callback,
+              Just row <- typedCb indexedItems =
+                -- set the fixed index to the target row and let the viewport position be sorted out by the
+                -- adjustment that follows the addition and resizing of the rows around the target row.
+                let newModel =
+                      cpModel
+                        { fixedRowIndex = row,
+                          fixedRowViewportOffset = 0,
+                          itemsBeforeInflated = row,
+                          inflatedItems = mempty,
+                          phase = ContentPaneReinflating
+                        }
+                 in [Model newModel, resizeInnerRequest wenv]
+            | otherwise =
+                []
+
+          indexedItems =
+            model.sortedItems
+              & S.mapWithIndex (,)
+              & S.sortOn (snd . snd)
+              & fmap (\(sortedIndex, (item, _originalIndex)) -> (item, sortedIndex))
+      InnerResizeComplete -> result
+        where
+          (rowsStartY, rowHeights, rowsEndY) = rowPositions node
+
+          fixedRowY = rowsStartY + sum (S.take (cpModel.fixedRowIndex - cpModel.itemsBeforeInflated) rowHeights)
+
+          result
+            | itemsToPrepend > 0 || itemsToAppend > 0 =
+                let inflatedItems =
+                      takeAt
+                        (cpModel.itemsBeforeInflated - itemsToPrepend)
+                        (itemsToPrepend + length cpModel.inflatedItems + itemsToAppend)
+                        model.sortedItems
+                    itemsBeforeInflated = cpModel.itemsBeforeInflated - itemsToPrepend
+                 in [Model cpModel {inflatedItems, itemsBeforeInflated}, resizeInnerRequest wenv]
+            | otherwise =
+                -- Once we have finished adding items then, if the added items are not the same size as estimated,
+                -- the row we want to scroll to might no longer be at the correct position in the viewport, so we
+                -- need to adjust the scroll position to position it correctly.
+                let adjustScrollEvt = [Report (ScrollToRect adjustScrollRect) | needAdjustScroll]
+                    adjustScrollRect = Rect (vp._rX + cpModel.visibleArea._rX) (vp._rY + fixedRowY - cpModel.fixedRowViewportOffset) visibleWidth visibleHeight
+                    needAdjustScroll = abs ((fixedRowY - cpModel.visibleArea._rY) - cpModel.fixedRowViewportOffset) >= 1
+                 in [Model cpModel {phase = ContentPaneIdle}] <> adjustScrollEvt
+
+          itemsToPrepend = itemsToAdd (fixedRowY - rowsStartY) 1 cpModel.itemsBeforeInflated
+          itemsToAppend = itemsToAdd (rowsEndY - fixedRowY) 2 (itemsAfterInflated model cpModel)
+
+          itemsToAdd existingItemsHeight f availableItems
+            | not (null model.sortedItems) && existingItemsHeight < visibleHeight * f =
+                let n = ceiling ((visibleHeight * f - existingItemsHeight) / model.mdlEstimatedItemHeight)
+                 in min (max 8 (min 64 n)) availableItems
+            | otherwise = 0
+
+          vp = node ^. L.info . L.viewport
+          visibleWidth = cpModel.visibleArea._rW
+          visibleHeight = cpModel.visibleArea._rH
+      ContentPaneParentEvent ep ->
+        [Report (ParentEvent ep)]
+
+contentPaneInner ::
+  forall a ep.
+  (CompositeModel a, WidgetEvent ep) =>
+  Seq (Column ep a) ->
+  HagridModel a ->
+  ContentPaneModel a ->
+  WidgetNode (ContentPaneModel a) (ContentPaneEvent ep)
+contentPaneInner columnDefs model cpModel = node
   where
     node =
-      defaultWidgetNode "Hagrid.ContentPane" contentPaneContainer
-        & L.children .~ S.fromList (mconcat childWidgetRows)
-
-    childWidgetRows =
-      [ [cellWidget idx item widget | Column {widget} <- columnDefs]
-        | (item, idx) <- model.sortedItems
-      ]
-
-    nCols = length columnDefs
-    columnDefsSeq = S.fromList columnDefs
+      defaultWidgetNode "Hagrid.ContentPaneInner" contentPaneContainer
+        & L.children .~ rowWidgets
 
     contentPaneContainer =
       createContainer
-        model
+        cpModel
         def
-          { containerGetSizeReq = getSizeReq,
-            containerResize = resize,
-            containerRender = render,
-            containerHandleEvent = handleEvent,
-            containerHandleMessage = handleMessage
+          { containerMerge = merge,
+            containerGetSizeReq = getSizeReq,
+            containerResize = resize
           }
+
+    rowWidgets = contentPaneRow columnDefs cpModel <$> cpModel.inflatedItems
+
+    merge _wenv newNode _oldNode oldState = resultReqs newNode reqs
+      where
+        reqs = [ResizeWidgets (newNode ^. L.info . L.widgetId) | needResize]
+        needResize = oldState.columnWidths /= cpModel.columnWidths
 
     getSizeReq _wenv _node children = (w, h)
       where
-        w = fixedSize (sum (fromIntegral . currentWidth <$> model.columns))
-        h = fixedSize (sum (toRowHeights children columnDefsSeq))
+        w = fixedSize (fromIntegral (sum cpModel.columnWidths))
+        h = fixedSize (uninflatedHeights + inflatedHeights)
 
-    resize wenv node viewport children = (resultNode node, assignedAreas)
+        uninflatedHeights = fromIntegral uninflatedItems * model.mdlEstimatedItemHeight
+        uninflatedItems = cpModel.itemsBeforeInflated + itemsAfterInflated model cpModel
+
+        inflatedHeights = sum (_szrFixed . _wniSizeReqH . _wnInfo <$> children)
+
+    resize wenv node viewport children = (resultEvts node [InnerResizeComplete], rowAreas)
       where
         style = currentStyle wenv node
-        Rect l t _w _h = fromMaybe def (removeOuterBounds style viewport)
+        innerVp = fromMaybe def (removeOuterBounds style viewport)
 
-        colXs = sizesToPositions (S.fromList (fromIntegral . currentWidth <$> model.columns))
-        rowYs = sizesToPositions (toRowHeights children columnDefsSeq)
+        startX = innerVp._rX
+        startY = innerVp._rY + fromIntegral cpModel.itemsBeforeInflated * model.mdlEstimatedItemHeight
 
-        assignedAreas = do
-          (rowN, row) <-
-            S.mapWithIndex (,) (S.chunksOf nCols children)
-          (colN, columnDef, widget) <-
-            S.mapWithIndex (\i (cd, w) -> (i, cd, w)) (S.zip columnDefsSeq row)
-          pure (assignArea colN columnDef rowN widget)
+        sumColumnWidths = fromIntegral (sum cpModel.columnWidths)
 
-        assignArea col Column {paddingW, paddingH, align} row widget = Rect chX chY chW chH
+        rowAreas = snd (foldl' foldRowAreas (startY, mempty) children)
+        foldRowAreas (y, areas) child =
+          (y + h, areas :|> Rect startX y sumColumnWidths h)
           where
-            (chX, chW)
-              | widgetReqW >= cellW = (cellX, cellW)
-              | align == ColumnAlignLeft = (cellX, widgetReqW)
-              | otherwise = (cellX + cellW - widgetReqW, widgetReqW)
-            (chY, chH) =
-              (cellY, cellH)
+            h = child ^. L.info . L.sizeReqH . L.fixed
 
-            cellX = l + S.index colXs col + paddingW
-            cellY = t + S.index rowYs row + paddingH
-            cellW = S.index colXs (col + 1) - S.index colXs col - paddingW * 2
-            cellH = S.index rowYs (row + 1) - S.index rowYs row - paddingH * 2
+contentPaneRow ::
+  forall a ep.
+  (CompositeModel a, WidgetEvent ep) =>
+  Seq (Column ep a) ->
+  ContentPaneModel a ->
+  (a, Int) ->
+  WidgetNode (ContentPaneModel a) (ContentPaneEvent ep)
+contentPaneRow columnDefs cpModel (item, rowIdx) = tree
+  where
+    tree =
+      defaultWidgetNode "Hagrid.Row" widget
+        & L.children .~ cellWidgets
 
-            widgetReqW =
+    widget =
+      createContainer
+        (cpModel.columnWidths, item, rowIdx)
+        def
+          { containerGetSizeReq = getSizeReq,
+            containerResize = resize,
+            containerRender = render
+          }
+
+    cellWidgets = do
+      Column {widget} <- columnDefs
+      pure (cellWidget rowIdx item widget)
+
+    getSizeReq _wenv _node children = (w, h)
+      where
+        w = fixedSize (fromIntegral (sum cpModel.columnWidths))
+        h = fixedSize (toRowHeight columnDefs children)
+
+    resize wenv node viewport children = (resultNode node, cellAreas)
+      where
+        style = currentStyle wenv node
+        innerVp = fromMaybe def (removeOuterBounds style viewport)
+
+        startX = innerVp._rX
+        startY = innerVp._rY
+
+        columnWidths = fromIntegral <$> S.fromList cpModel.columnWidths
+
+        cellAreas = snd (foldl' foldCellAreas (startX, mempty) (S.zip3 columnWidths columnDefs children))
+        foldCellAreas (x, areas) (colW, Column {paddingW, paddingH, align}, widget) =
+          (x + colW, areas :|> Rect chX cellY chW cellH)
+          where
+            (chX, chW) = case align of
+              ColumnAlignLeft -> (cellX, cellW)
+              ColumnAlignRight -> (cellX + cellW - widgetW, widgetW)
+
+            cellX = x + paddingW
+            cellY = startY + paddingH
+            cellW = colW - paddingW * 2
+            cellH = viewport._rH - paddingH * 2
+
+            widgetW =
               widget
                 & _wnInfo
                 & _wniSizeReqW
-                & \r -> _szrFixed r + _szrFlex r
+                & (\r -> _szrFixed r + _szrFlex r)
+                & min cellW
 
     render wenv node renderer = do
-      forM_ (neighbours rowYs) $ \(y1, y2, even) -> do
-        let color
-              | mouseover && _pY mouse >= (t + y1) && _pY mouse < (t + y2) = Just mouseOverColor
-              | not even = Just oddRowBgColor
-              | otherwise = Nothing
-        drawRect renderer (Rect l (t + y1) lastColX (y2 - y1)) color Nothing
-
-      forM_ (S.drop 1 colXs) $ \colX -> do
-        drawLine renderer (Point (l + colX) t) (Point (l + colX) (t + lastRowY)) 1 (Just lineColor)
-
-      forM_ (S.drop 1 rowYs) $ \rowY -> do
-        drawLine renderer (Point l (t + rowY)) (Point (l + lastColX) (t + rowY)) 1 (Just lineColor)
+      drawRect renderer vp bgColor Nothing
+      drawLine renderer (Point vp._rX (vp._rY + vp._rH)) (Point (vp._rX + vp._rW) (vp._rY + vp._rH)) 1 (Just lineColor)
+      forM_ (drop 1 colXs) $ \colX -> do
+        drawLine renderer (Point (vp._rX + colX) vp._rY) (Point (vp._rX + colX) (vp._rY + vp._rH)) 1 (Just lineColor)
       where
-        colXs = sizesToPositions (S.fromList (fromIntegral . currentWidth <$> model.columns))
-        rowYs = sizesToPositions (toRowHeights (node ^. L.children) columnDefsSeq)
-        lastColX
-          | _ :|> a <- colXs = a
-          | otherwise = 0
-        lastRowY
-          | _ :|> a <- rowYs = a
-          | otherwise = 0
+        colXs = scanl (+) 0 (fromIntegral <$> cpModel.columnWidths)
+        bgColor
+          | mouseover = Just mouseOverColor
+          | rowIdx `mod` 2 == 1 = Just oddRowBgColor
+          | otherwise = Nothing
         vp = node ^. L.info . L.viewport
-        Rect l t _w _h = vp
         mouseover = pointInRect mouse vp
         mouse = wenv ^. L.inputStatus . L.mousePos
         mouseOverColor = (accentColor wenv) {_colorA = 0.3}
         oddRowBgColor = (accentColor wenv) {_colorA = 0.1}
         lineColor = accentColor wenv
 
-    handleEvent _wenv node _path = \case
-      Move (Point _pX _pY) ->
-        -- refresh which row shows as hovered
-        Just (resultReqs node [RenderOnce])
-      _ -> Nothing
-
-    handleMessage :: ContainerMessageHandler (HagridModel a) (HagridEvent ep)
-    handleMessage wenv node _path msg = result
-      where
-        result = cast msg >>= handleTypedMessage
-
-        handleTypedMessage (ContentPaneScrollToRow callback) = result
-          where
-            result
-              | Just row <- callback indexedItems,
-                Just y1 <- S.lookup row rowYs,
-                Just y2 <- S.lookup (row + 1) rowYs =
-                  Just (resultReqs node [SendMessage scrollId (ScrollTo (Rect vp._rX (vp._rY + y1) 1 (y2 - y1)))])
-              | otherwise =
-                  Nothing
-
-            indexedItems =
-              model.sortedItems
-                & indexed
-                & List.sortOn (snd . snd)
-                & map (\(sortedIndex, (item, _originalIndex)) -> (item, sortedIndex))
-
-            vp = node ^. L.info . L.viewport
-            rowYs = sizesToPositions (toRowHeights (node ^. L.children) columnDefsSeq)
-
-            scrollId = fromJust (widgetIdFromKey wenv (WidgetKey contentScrollKey))
-
-initialModel :: [HagridCfg s e] -> [Column ep a] -> [a] -> HagridModel a
-initialModel cfg columnDefs items = model
+initialModel :: [HagridCfg s e] -> [Column ep a] -> Seq a -> HagridModel a
+initialModel cfgs columnDefs items = model
   where
     model =
       HagridModel
-        { sortedItems = sortItems columnDefs sortColumn (zip items [0 ..]),
+        { sortedItems = sortItems columnDefs sortColumn (S.zip items (S.fromFunction (length items) id)),
           columns = initialColumn <$> columnDefs,
-          sortColumn
+          sortColumn,
+          mdlEstimatedItemHeight = max 1 (fromMaybe 40 cfg.cfgEstimatedItemHeight)
         }
 
+    cfg = mconcat cfgs
+
     sortColumn
-      | Just (col, dir) <- (mconcat cfg).cfgInitialSort,
+      | Just (col, dir) <- cfg.cfgInitialSort,
         col >= 0,
         col < length columnDefs =
           Just (col, dir)
       | otherwise = Nothing
 
-    initialColumn Column {name, initialWidth} =
+    initialColumn Column {name, initialWidth, minWidth} =
       ModelColumn
         { name,
-          currentWidth = initialWidth
+          currentWidth = max minWidth initialWidth
         }
+
+-- | When the viewport position changes, this function computes the index and
+-- position within the viewport of the new fixed row. This is the first row
+-- whose y-position is at least minVisibleY
+fixedRow :: Double -> Seq Double -> HagridModel a -> ContentPaneModel a -> (Int, Double)
+fixedRow minVisibleY inflatedItemHeights model cpModel = (min maxRow row, offset)
+  where
+    (row, offset)
+      | minVisibleY < inflatedStartY =
+          let row = ceiling (minVisibleY / model.mdlEstimatedItemHeight)
+              offset = (fromIntegral row * model.mdlEstimatedItemHeight) - minVisibleY
+           in (row, offset)
+      | otherwise =
+          inflatedItem itemsBeforeInflated inflatedStartY inflatedItemHeights
+
+    maxRow = length model.sortedItems - 1
+
+    ContentPaneModel {itemsBeforeInflated} = cpModel
+
+    inflatedStartY = fromIntegral itemsBeforeInflated * model.mdlEstimatedItemHeight
+
+    inflatedItem i y = \case
+      itemHeight :<| itemHeights
+        | y + itemHeight >= minVisibleY ->
+            (i + 1, (y + itemHeight) - minVisibleY)
+        | otherwise ->
+            inflatedItem (i + 1) (y + itemHeight) itemHeights
+      S.Empty ->
+        let indexInSection = ceiling ((minVisibleY - y) / model.mdlEstimatedItemHeight)
+            row = i + indexInSection
+            offset = (y + fromIntegral indexInSection * model.mdlEstimatedItemHeight) - minVisibleY
+         in (row, offset)
+
+resizeInnerRequest :: WidgetEnv s e -> EventResponse s e sp ep
+resizeInnerRequest wenv =
+  Request (ResizeWidgets (fromJust (widgetIdFromKey wenv (WidgetKey contentPaneInnerKey))))
+
+itemsAfterInflated :: HagridModel a -> ContentPaneModel a -> Int
+itemsAfterInflated model cpModel =
+  length model.sortedItems - (cpModel.itemsBeforeInflated + length cpModel.inflatedItems)
 
 dragHandleWidth :: Int
 dragHandleWidth = 4
@@ -730,8 +962,11 @@ headerPaneKey = "Hagrid.headerPane"
 contentScrollKey :: Text
 contentScrollKey = "Hagrid.contentScroll"
 
-contentPaneKey :: Text
-contentPaneKey = "Hagrid.contentPane"
+contentPaneOuterKey :: Text
+contentPaneOuterKey = "Hagrid.contentPaneOuter"
+
+contentPaneInnerKey :: Text
+contentPaneInnerKey = "Hagrid.contentPaneInner"
 
 footerPaneKey :: Text
 footerPaneKey = "Hagrid.footerPane"
@@ -739,12 +974,12 @@ footerPaneKey = "Hagrid.footerPane"
 sortItems ::
   [Column ep a] ->
   Maybe (Int, SortDirection) ->
-  [ItemWithIndex a] ->
-  [ItemWithIndex a]
+  Seq (ItemWithIndex a) ->
+  Seq (ItemWithIndex a)
 sortItems columnDefs sortColumn items =
   case modelSortKey columnDefs sortColumn of
     DontSort -> items
-    SortWith f -> List.sortOn (f . fst) items
+    SortWith f -> S.sortOn (f . fst) items
 
 modelSortKey :: [Column ep a] -> Maybe (Int, SortDirection) -> ColumnSortKey a
 modelSortKey columnDefs sortColumn = case sortColumn of
@@ -756,11 +991,8 @@ modelSortKey columnDefs sortColumn = case sortColumn of
   _ ->
     DontSort
 
-sizesToPositions :: Seq Double -> Seq Double
-sizesToPositions = S.scanl (+) 0
-
-toRowHeights :: Seq (WidgetNode s e1) -> Seq (Column e2 a) -> Seq Double
-toRowHeights children columnDefs = mergeHeights <$> S.chunksOf (length columnDefs) children
+toRowHeight :: Seq (Column e2 a) -> Seq (WidgetNode s e1) -> Double
+toRowHeight columnDefs = mergeHeights
   where
     mergeHeights rowWidgets =
       foldl' max 0 (S.zipWith widgetHeight columnDefs rowWidgets)
@@ -771,11 +1003,23 @@ toRowHeights children columnDefs = mergeHeights <$> S.chunksOf (length columnDef
         & _wniSizeReqH
         & \r -> _szrFixed r + _szrFlex r + paddingH * 2
 
-neighbours :: Seq a -> Seq (a, a, Bool)
-neighbours = \case
-  a :<| b :<| c :<| rest -> (a, b, False) :<| (b, c, True) :<| neighbours (c :<| rest)
-  a :<| b :<| S.Empty -> S.singleton (a, b, False)
-  _ -> S.empty
+rowPositions :: forall s e. WidgetNode s e -> (Double, Seq Double, Double)
+rowPositions node = (rowsStartY, rowHeights, rowsEndY)
+  where
+    vp = node._wnInfo._wniViewport
+    childVps = _wniViewport . _wnInfo <$> node._wnChildren
+
+    rowsStartY = case childVps of
+      cvp :<| _ -> cvp._rY - vp._rY
+      _ -> 0
+    rowsEndY = case childVps of
+      _ :|> cvp -> (cvp._rY - vp._rY) + cvp._rH
+      _ -> 0
+    rowHeights = _rH <$> childVps
+
+takeAt :: Int -> Int -> Seq a -> Seq a
+takeAt at len s =
+  S.take len (S.drop at s)
 
 -- | Creates a column that displays a text value, and is sortable by the text.
 textColumn ::
@@ -828,7 +1072,12 @@ defaultColumn name widget =
       sortHandler = Nothing
     }
 
-cellWidget :: (CompositeModel a, WidgetEvent e, WidgetModel s) => Int -> a -> ColumnWidget e a -> WidgetNode (HagridModel s) (HagridEvent e)
+cellWidget ::
+  (CompositeModel a, WidgetEvent e, WidgetModel s) =>
+  Int ->
+  a ->
+  ColumnWidget e a ->
+  WidgetNode (ContentPaneModel s) (ContentPaneEvent e)
 cellWidget idx item = \case
   LabelWidget get -> label_ (get idx item) [ellipsis]
   CustomWidget get -> widget
@@ -838,11 +1087,11 @@ cellWidget idx item = \case
       buildUI _wenv =
         get idx
       handleEvent _wenv _node _model e =
-        [Report (ParentEvent e)]
+        [Report (ContentPaneParentEvent e)]
 
 footerWidgetNode ::
   (CompositeModel a, CompositeModel s, Typeable e) =>
-  [ItemWithIndex a] ->
+  Seq (ItemWithIndex a) ->
   ColumnFooterWidget e a ->
   Maybe (WidgetNode (HagridModel s) (HagridEvent e))
 footerWidgetNode items = \case
@@ -858,6 +1107,10 @@ footerWidgetNode items = \case
 
 -- | Sends a message to the targeted 'hagrid' widget, that causes the
 -- widget to scroll such that a specified row becomes visible.
+--
+-- Note that this is inherently dynamically typed. If the type of the callback
+-- does not match the type of the targeted hagrid widget then the message
+-- will be ignored.
 scrollToRow ::
   forall s e sp ep a.
   (Typeable a, Typeable e) =>
@@ -884,3 +1137,12 @@ flipSortDirection SortDescending = SortAscending
 
 ceilingDouble :: Double -> Double
 ceilingDouble x = fromIntegral (ceiling x :: Int)
+
+roundedRectEq :: Rect -> Rect -> Bool
+roundedRectEq r1 r2 =
+  roundedEq r1._rX r2._rX
+    && roundedEq r1._rY r2._rY
+    && roundedEq (r1._rX + r1._rW) (r2._rX + r2._rW)
+    && roundedEq (r1._rY + r1._rH) (r2._rY + r2._rH)
+  where
+    roundedEq x y = (round x :: Int) == round y