+++ ChangeLog.md
@@ -1,6 +1,8 @@
-## 0.1.0.1
+## 0.2.0.0
 ### Added
-- New "resizing-cells" example to show what happens when cells resize. This will be useful to demonstrate that virtual grids work sensibly.
+- Speed up large grids by only creating widgets that are visible.
+- Use `Seq` rather than lists, for better performance with large grids (breaks API).
+- New "resizing-cells" example to show what happens when cells resize.
 ### Fixed
 - The footer now only takes up vertical space if there are some footer widgets defined.